rpc, ethclient: Add Transaction type corresponding to RPC documention.
The docs here (github.com/ethereum/wiki/wiki/JSON-RPC) define a schema for transactions returned over RPC that is different from the `types.Transaction` structure, which is the current return value for RPC methods in the built-in Go RPC client.

This adds a new type, `rpc.Transaction` that matches the documentation, and updates `ethclient.TransactionByHash` to use that return type.